refactor(frontend-v2): address chapters PR review feedback
- Show an explicit error notice when the chapter list query fails
- Encode the Gmail compose link via URLSearchParams and open it with
  noopener,noreferrer
- Fix the "SuriName" display typo (safe: the persisted value is the
  country code, not the name)
- Reject non-numeric chapter IDs strictly instead of via parseInt
- Gate /chapters/new with the session role check (admin or
  intl_coordinator, mirroring apiIntlCoordinatorAuthMiddleware) instead
  of a throwaway fetch
- Use zod's built-in email validator
- Split the chapter form into schema/hook/section modules with a
  buildChapterPayload helper
- Use date-fns for the date helpers in chapter-utils
