+++ frontend-v2/src/app/(authed)/chapters/chapter-utils.ts
@@ -1,3 +1,18 @@
+import {
+  addDays,
+  addHours,
+  addMonths,
+  differenceInCalendarDays,
+  format,
+  isAfter,
+  isBefore,
+  isValid,
+  parse,
+  subDays,
+  subHours,
+  subMonths,
+} from 'date-fns'
+
 export const REGIONS = [
   'North America',
   'Central & South America',
@@ -11,29 +26,22 @@ export type StatusColor = 'green' | 'yellow' | 'red' | 'gray' | 'black'
 
 /** Parses a `YYYY-MM-DD` string as a local calendar date (no timezone shift). */
 export function parseDateYmd(value: string): Date | null {
-  const match = /^(\d{4})-(\d{2})-(\d{2})$/.exec(value)
-  if (!match) return null
-  const [, year, month, day] = match
-  return new Date(Number(year), Number(month) - 1, Number(day))
+  const parsed = parse(value, 'yyyy-MM-dd', new Date())
+  return isValid(parsed) ? parsed : null
 }
 
 export function formatDateYmd(date: Date): string {
-  const year = date.getFullYear()
-  const month = String(date.getMonth() + 1).padStart(2, '0')
-  const day = String(date.getDate()).padStart(2, '0')
-  return `${year}-${month}-${day}`
+  return format(date, 'yyyy-MM-dd')
 }
 
 export function isDateInLastThreeMonths(value: string): boolean {
   const date = parseDateYmd(value)
   if (date == null) return false
-  const threeMonthsAgo = new Date()
-  threeMonthsAgo.setMonth(threeMonthsAgo.getMonth() - 3)
-  return date > threeMonthsAgo
+  return isAfter(date, subMonths(new Date(), 3))
 }
 
 // Quadrimesters: Feb-May, Jun-Sep, Oct-Jan. Returns the first day of the
-// quadrimester containing today.
+// quadrimester containing today. Ported from frontend/ChapterList.vue.
 function currentQuadrimesterStart(now: Date): Date {
   const month = now.getMonth()
   const year = now.getFullYear()
@@ -43,47 +51,35 @@ function currentQuadrimesterStart(now: Date): Date {
   return new Date(year - 1, 9, 1) // January -> previous Oct
 }
 
-function addMonths(date: Date, months: number): Date {
-  return new Date(date.getFullYear(), date.getMonth() + months, date.getDate())
-}
-
-function addDays(date: Date, days: number): Date {
-  const result = new Date(date)
-  result.setDate(result.getDate() + days)
-  return result
-}
-
 export function colorLastAction(value: string): StatusColor {
   const now = new Date()
   const quadStart = currentQuadrimesterStart(now)
-  const prevQuadStart = addMonths(quadStart, -4)
+  const prevQuadStart = subMonths(quadStart, 4)
   const quadEnd = addMonths(quadStart, 4)
   const redThreshold = addDays(addMonths(quadStart, 1), 14)
-  const blackThreshold = addDays(quadEnd, -7)
+  const blackThreshold = subDays(quadEnd, 7)
 
   const lastAction = parseDateYmd(value)
-  if (lastAction == null || lastAction < prevQuadStart) return 'black'
+  if (lastAction == null || isBefore(lastAction, prevQuadStart)) return 'black'
 
-  const hasActionThisQuadrimester = lastAction >= quadStart
+  const hasActionThisQuadrimester = !isBefore(lastAction, quadStart)
   if (hasActionThisQuadrimester) return 'green'
-  if (now < redThreshold) return 'green'
-  if (now < blackThreshold) return 'red'
+  if (isBefore(now, redThreshold)) return 'green'
+  if (isBefore(now, blackThreshold)) return 'red'
   return 'black'
 }
 
 export function lastActionTooltip(value: string): string {
   const now = new Date()
   const quadStart = currentQuadrimesterStart(now)
-  const quadEnd = addDays(addMonths(quadStart, 4), -1)
+  const quadEnd = subDays(addMonths(quadStart, 4), 1)
   const quadrimesterText = `Current quadrimester: ${formatDateYmd(quadStart)} to ${formatDateYmd(quadEnd)}`
 
   const lastAction = parseDateYmd(value)
   if (lastAction == null) {
     return `No action recorded\n${quadrimesterText}`
   }
-  const daysSinceLastAction = Math.floor(
-    (now.getTime() - lastAction.getTime()) / (1000 * 60 * 60 * 24),
-  )
+  const daysSinceLastAction = differenceInCalendarDays(now, lastAction)
   return `${daysSinceLastAction} days since last action\n${quadrimesterText}`
 }
 
@@ -93,19 +89,15 @@ export function colorFBSyncStatus(value: string): StatusColor {
   if (date == null) return 'gray'
 
   const now = new Date()
-  const oneHourAgo = new Date(now.getTime() - 60 * 60 * 1000)
-  const oneDayAgo = new Date(now.getTime() - 24 * 60 * 60 * 1000)
-
-  if (date > oneHourAgo) return 'green'
-  if (date > oneDayAgo) return 'yellow'
+  if (isAfter(date, subHours(now, 1))) return 'green'
+  if (isAfter(date, subDays(now, 1))) return 'yellow'
   return 'red'
 }
 
 function parseFBSyncDate(value: string): Date | null {
   if (!value) return null
   const parsed = new Date(value)
-  if (Number.isNaN(parsed.getTime())) return null
-  return new Date(parsed.getTime() + 8 * 60 * 60 * 1000)
+  return isValid(parsed) ? addHours(parsed, 8) : null
 }
 
 /** Builds a Gmail compose link addressed to a chapter's own email plus all of its organizers' emails. */
@@ -119,7 +111,13 @@ export function buildChapterEmailLink(chapter: {
     ...chapter.Organizers.map((o) => o.Email),
   ].filter(Boolean)
   if (emails.length === 0) return null
-  return `https://mail.google.com/mail/?view=cm&fs=1&su=${chapter.Name}&to=${emails.join(',')}`
+  const params = new URLSearchParams({
+    view: 'cm',
+    fs: '1',
+    su: chapter.Name,
+    to: emails.join(','),
+  })
+  return `https://mail.google.com/mail/?${params}`
 }
 
 export const STATUS_COLOR_CLASSES: Record<StatusColor, string> = {
