+++ frontend-v2/src/app/(authed)/chapters/new/page.tsx
@@ -1,15 +1,15 @@
+import { forbidden } from 'next/navigation'
 import { ContentWrapper } from '@/app/content-wrapper'
-import { ApiClient } from '@/lib/api'
-import { getCookies } from '@/lib/auth'
-import { redirectForHttpError } from '@/lib/server-auth'
+import { getCachedSession } from '@/app/session'
 import { ChapterForm } from '../chapter-form'
 
 export default async function NewChapterPage() {
-  const apiClient = new ApiClient(await getCookies())
-
-  // Not used by the form itself; this call exists purely to gate the page
-  // behind the same admin/intl_coordinator check as the rest of /chapters.
-  await redirectForHttpError(() => apiClient.getChapterAdminList())
+  const session = await getCachedSession()
+  // Mirrors the Go apiIntlCoordinatorAuthMiddleware role check.
+  const roles = session.user?.Roles ?? []
+  if (!roles.includes('admin') && !roles.includes('intl_coordinator')) {
+    forbidden()
+  }
 
   return (
     <ContentWrapper size="lg" className="gap-6">
